Smart Alarm App
Smart Alarm App: When alarm rings puzzle appears on your screen.Puzzle
involves arranging 2D tiles of dimension 2X2 to complete a picture. It
randomly chooses a image from the set of images in the internal list
and then it disarranges image and ask you to solve the puzzle. The stop
button is disabled until you solve the puzzle.